slapd casi acabado

+++ services/purge_all.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 
-docker-compose stop
-docker rm $(docker-compose ps -q)
-docker images rm $(docker-compose images -q)
+COMPOSEFILE=$1
+
+if [[ ! -z $COMPOSEFILE ]]; then
+    [[ -f $COMPOSEFILE ]] && ARGS="-f $COMPOSEFILE" || exit 1
+fi
+
+docker-compose $ARGS stop
+docker rm $(docker-compose $ARGS ps -q)
+docker images rm $(docker-compose $ARGS images -q)
 docker volume rm $(docker volume ls -q)
 docker network prune
